+++ app/src/main/java/com/android/example/tanmen/View/BottomNavigationPagerAdapter.kt
@@ -1,12 +1,11 @@
 package com.android.example.tanmen.View
 
 import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
 import androidx.viewpager2.adapter.FragmentStateAdapter
 import com.android.example.tanmen.Controller.Fragment.MainFragment
 import com.android.example.tanmen.Controller.Fragment.ShuffleFragment
 
-class BottomNavigationPagerAdapter(fm : FragmentActivity) : FragmentStateAdapter(fm) {
+class BottomNavigationPagerAdapter(fm: MainFragment) : FragmentStateAdapter(fm) {
     override fun getItemCount(): Int {
         return 2
     }
